main.py
import os
import cv2
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import logging
import warnings
from PIL import Image
from sklearn.cluster import DBSCAN
from sklearn.model_selection import KFold, cross_val_score, GridSearchCV, train_test_split
from sklearn.metrics import mean_absolute_error, mean_squared_error
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestRegressor, StackingRegressor
from xgboost import XGBRegressor
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from sklearn.base import BaseEstimator, RegressorMixin
import tensorflow as tf
from tensorflow.keras.models import Model, Sequential
from tensorflow.keras.layers import (
    Conv2D, MaxPooling2D, Flatten, Dense, Dropout,
    Input, BatchNormalization, Activation, concatenate
)
from tensorflow.keras.optimizers import Adam
from tensorflow.keras.applications import VGG16, ResNet50, EfficientNetB0
from tensorflow.keras.callbacks import EarlyStopping, ModelCheckpoint
from tensorflow.keras.preprocessing.image import ImageDataGenerator
from tensorflow.keras.utils import plot_model
from tensorflow.keras import backend as K

try:
    from skimage import feature
except ModuleNotFoundError:
    print("scikit-image module not found. Please install it using '!pip install --user scikit-image'")

# Set logging
logging.basicConfig(level=logging.INFO)
warnings.filterwarnings('ignore')

# Constants
IMAGE_DIR = './documents'
PREPROCESSED_DIR = './documents/preprocessed_images'
EXCEL_FILE = 'Meta_pic_3.xlsx'
LEFT, TOP, IMG_WIDTH, IMG_HEIGHT = 232, 60, 495, 475
RIGHT, BOTTOM = LEFT + IMG_WIDTH, TOP + IMG_HEIGHT
CROP_TOP, CROP_HEIGHT, CROP_LEFT, CROP_WIDTH = 290, 100, 215, 125
CROP_BOTTOM, CROP_RIGHT = CROP_TOP + CROP_HEIGHT, CROP_LEFT + CROP_WIDTH
MAX_FEATURES, GOOD_MATCH_PERCENT, MAX_DELTA = 500, 0.15, 200
MIN_DELTA_X, MAX_DELTA_X = -CROP_LEFT, IMG_WIDTH - CROP_RIGHT
MIN_DELTA_Y, MAX_DELTA_Y = -CROP_TOP, IMG_HEIGHT - CROP_BOTTOM
NUM_FOLDS = 5  # For K-fold cross-validation

# Create directories
os.makedirs(PREPROCESSED_DIR, exist_ok=True)

# Configure TensorFlow to avoid OOM errors
gpus = tf.config.experimental.list_physical_devices('GPU')
if gpus:
    try:
        for gpu in gpus:
            tf.config.experimental.set_memory_growth(gpu, True)
        logging.info("GPU memory growth set")
    except RuntimeError as e:
        logging.warning(e)

def load_and_preprocess_images(image_dir):
    # Function to load and preprocess images
    pass

def align_images(reference_image, images):
    # Function for image alignment using ORB
    pass

def preprocess_and_save_images(images, offsets, preprocessed_dir):
    # Function to preprocess images (cropping and saving)
    pass

def augment_images(images, labels):
    # Function to apply image augmentation
    pass

def load_labels_and_images(excel_file, preprocessed_dir):
    # Function to load labels and match with images
    pass

def build_unet_model(input_shape=(128, 128, 1)):
    # Function to build UNet model
    pass

def cross_validate_model(model_fn, X, y, n_splits=NUM_FOLDS, epochs=25, batch_size=32):
    # Function to perform K-fold cross-validation
    pass

def train_ensemble_models(X_train_flat, y_train):
    # Function to build and train ensemble models
    pass

def tune_hyperparameters(X_train_flat, y_train):
    # Function to perform hyperparameter tuning
    pass

def main():
    # Main execution flow
    logging.info("Loading and preprocessing images...")
    raw_images = load_and_preprocess_images(IMAGE_DIR)
    reference_image = raw_images[0][1]
    aligned_images, offsets = align_images(reference_image, raw_images)
    preprocess_and_save_images(aligned_images, offsets, PREPROCESSED_DIR)

    logging.info("Loading labels and images...")
    images, labels = load_labels_and_images(EXCEL_FILE, PREPROCESSED_DIR)

    logging.info("Applying edge detection...")
    images_edges = np.array([feature.canny(img.astype('float32') / 255.0) for img in images])
    images_edges = images_edges[..., np.newaxis]

    images = images.astype('float32') / 255.0
    images = images[..., np.newaxis]

    logging.info("Augmenting images...")
    images_augmented, labels_augmented = augment_images(images, labels)
    images_combined = np.vstack((images, images_augmented))
    labels_combined = np.hstack((labels, labels_augmented))

    logging.info("Splitting data...")
    X_train, X_test, y_train, y_test = train_test_split(images_combined, labels_combined, test_size=0.2, random_state=42)

    logging.info("Building and cross-validating UNet model...")
    unet_mae = cross_validate_model(build_unet_model, X_train, y_train)

    logging.info("Preparing data for traditional ML models...")
    X_train_flat = X_train.reshape(X_train.shape[0], -1)
    X_test_flat = X_test.reshape(X_test.shape[0], -1)

    logging.info("Hyperparameter tuning for XGBoost...")
    xgb_best = tune_hyperparameters(X_train_flat, y_train)

    logging.info("Training ensemble model...")
    stack_model = train_ensemble_models(X_train_flat, y_train)

    logging.info("Evaluating models...")
    unet_model = build_unet_model()
    unet_model.compile(optimizer=Adam(), loss='mse', metrics=['mae'])
    unet_model.fit(X_train, y_train, epochs=25, batch_size=32, validation_data=(X_test, y_test), verbose=0)
    unet_predictions = unet_model.predict(X_test).flatten()
    unet_mae = mean_absolute_error(y_test, unet_predictions)
    unet_rmse = np.sqrt(mean_squared_error(y_test, unet_predictions))

    xgb_predictions = xgb_best.predict(X_test_flat)
    xgb_mae = mean_absolute_error(y_test, xgb_predictions)
    xgb_rmse = np.sqrt(mean_squared_error(y_test, xgb_predictions))

    stack_predictions = stack_model.predict(X_test_flat)
    stack_mae = mean_absolute_error(y_test, stack_predictions)
    stack_rmse = np.sqrt(mean_squared_error(y_test, stack_predictions))

    logging.info("Compiling results...")
    comparison_df = pd.DataFrame({
        'Model': ['UNet', 'XGBoost (Tuned)', 'Ensemble (Stacked)'],
        'MAE': [unet_mae, xgb_mae, stack_mae],
        'RMSE': [unet_rmse, xgb_rmse, stack_rmse]
    })

    print(comparison_df)
    comparison_df.to_csv('optimized_model_comparison.csv', index=False)

    plt.figure(figsize=(10, 6))
    plt.bar(comparison_df['Model'], comparison_df['MAE'], color='skyblue')
    plt.xlabel('Model')
    plt.ylabel('Mean Absolute Error (MAE)')
    plt.title('Model Comparison - MAE')
    plt.savefig('optimized_model_mae_comparison.png')
    plt.show()

    def bland_altman_plot(actual, predicted, model_name):
        difference = actual - predicted
        avg = (actual + predicted) / 2
        mean_diff = np.mean(difference)
        std_diff = np.std(difference)
        upper_limit = mean_diff + 1.96 * std_diff
        lower_limit = mean_diff - 1.96 * std_diff

        plt.figure(figsize=(10, 6))
        plt.scatter(avg, difference, alpha=0.5)
        plt.axhline(mean_diff, color='red', linestyle='--', label='Mean Difference')
        plt.axhline(upper_limit, color='gray', linestyle='--', label='Upper Limit')
        plt.axhline(lower_limit, color='gray', linestyle='--', label='Lower Limit')
        plt.xlabel('Average of Actual and Predicted')
        plt.ylabel('Difference between Actual and Predicted')
        plt.title(f'Bland-Altman Plot - {model_name}')
        plt.legend()
        plt.savefig(f'bland_altman_{model_name.lower().replace(" ", "_")}.png')
        plt.show()

    bland_altman_plot(y_test, unet_predictions, 'UNet')

if __name__ == '__main__':
    main()    