requirements.txt
# requirements.txt content:
numpy>=1.21
pandas>=1.3
tensorflow>=2.17.0
scikit-learn>=1.0
matplotlib>=3.3
opencv-python-headless>=4.5
pillow>=8.2
xgboost>=1.4
jupyter>=1.0
Cython>=0.29
openpyxl>=3.0
wandb>=0.12
pingouin>=0.4
seaborn>=0.11
statsmodels>=0.12
scikit-plot>=0.3
imgaug
scikit-image
logging
warnings
numpy
pandas
matplotlib
scikit-learn
scikit-image
tensorflow
xgboost
opencv-python
